modified some executorHandler class

+++ linux-monitor-common/src/main/java/com/hotstrip/linux/monitor/common/pojo/SystemLoadAvgDO.java
@@ -1,11 +1,8 @@
 package com.hotstrip.linux.monitor.common.pojo;
 
 import lombok.Data;
-import lombok.EqualsAndHashCode;
 import lombok.NoArgsConstructor;
 import lombok.experimental.SuperBuilder;
-
-import java.math.BigDecimal;
 
 @Data
 @SuperBuilder
@@ -13,8 +10,8 @@ import java.math.BigDecimal;
 //@EqualsAndHashCode(callSuper = true)
 public class SystemLoadAvgDO {
 
-    private BigDecimal one;
-    private BigDecimal five;
-    private BigDecimal fifteen;
+    private Double one;
+    private Double five;
+    private Double fifteen;
 
 }
